data validation

+++ src/main/java/web/dao/UserDaoJDBCImpl.java
@@ -1,10 +1,10 @@
 package web.dao;
 
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 import web.model.User;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
@@ -15,32 +15,34 @@ public class UserDaoJDBCImpl implements UserDao {
     private EntityManager entityManager;
 
     @Override
-    @Transactional(readOnly = true)
     public List<User> getAll() {
         return entityManager.createQuery("select u from User u", User.class).getResultList();
     }
 
     @Override
-    @Transactional
     public void add(User user) {
         entityManager.persist(user);
     }
 
     @Override
-    @Transactional
     public void deleteById(Long id) {
         User user = entityManager.find(User.class, id);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id " + id + " does not exist.");
+        }
         entityManager.remove(user);
     }
 
     @Override
-    @Transactional(readOnly = true)
     public User getById(Long id) {
-        return entityManager.find(User.class, id);
+        User user = entityManager.find(User.class, id);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id " + id + " does not exist.");
+        }
+        return user;
     }
 
     @Override
-    @Transactional
     public void change(User user) {
         entityManager.merge(user);
     }
